requirements.txt
click
DBUtils
Flask
Flask-Cors
flask-session
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
PyMySQL
six
typing_extensions
zipp
pillow
apscheduler
loguru
requests
qrcode[pil]
redis
Werkzeug
pandas
PyJWT
pytz
paho-mqtt
cryptography
flask-sock
redis
rq
rq-scheduler
pycryptodome
tencentcloud-sdk-python-common
tencentcloud-sdk-python-ocr